Added know issue relating to Joyent Triton.
There's currently an issue relating to starting meteord containers on Joyent's Triton Infrastructure. I've been back and forth with Joyent's support team (Peter Gale and Andrew Hill) and they've come up with this workaround. They have a proper fix, but they told me it might take a while to roll out to production, so they gave me this work around to use in the meantime.

Once the fix has been rolled out, this can be removed, but there's currently no ETA for this to happen. I thought I'd share this here to save others some headaches.